requirements.txt
telethon
requests
pillow
pyzbar
